// packages/oceanengine-ad-open-sdk/src/models/ToolsWechatAppletListV30Filtering.ts
// Generated from oceanengine/ad_open_sdk_java
// Phase: B
// Do not edit manually.

import type { ToolsWechatAppletListV30FilteringAuditStatus, ToolsWechatAppletListV30FilteringCreateTime, ToolsWechatAppletListV30FilteringSearchType, ToolsWechatAppletListV30FilteringUpdateTime } from "../models";

export interface ToolsWechatAppletListV30Filtering {
  audit_status?: ToolsWechatAppletListV30FilteringAuditStatus;
  create_time?: ToolsWechatAppletListV30FilteringCreateTime;
  name?: string;
  search_type?: ToolsWechatAppletListV30FilteringSearchType;
  update_time?: ToolsWechatAppletListV30FilteringUpdateTime;
  user_name?: string;
}