// packages/oceanengine-ad-open-sdk/src/models/ToolsWechatAppletListV30Response.ts
// Generated from oceanengine/ad_open_sdk_java
// Phase: B
// Do not edit manually.

import type { ToolsWechatAppletListV30ResponseData } from "../models";

export interface ToolsWechatAppletListV30Response {
  code?: number;
  data?: ToolsWechatAppletListV30ResponseData;
  message?: string;
  request_id?: string;
}